+++ tablist.novaextension/CHANGELOG.md
@@ -1,8 +1,14 @@
+## Version 2.0
+
+- Create new grouped style view
+- New button to switch into views
+- Project configuration to select view
+
 ## Version 1.0.2
 
-Changed double click by item selection 
-Added on mouse over file path
+- Changed double click by item selection
+- Added on mouse over file path
 
 ## Version 1.0
 
-Initial release
+- Initial release